fix(nextjs-app-router-localisation-starter): fix next-intl errors during next build
