+++ examples/nextjs-app-router-localization-starter/package.json
@@ -15,7 +15,7 @@
     "@types/node": "20.10.6",
     "@types/react": "18.2.40",
     "@types/react-dom": "18.2.17",
-    "@uniformdev/canvas-next-rsc": "19.130.0",
+    "@uniformdev/canvas-next-rsc": "19.161.0",
     "autoprefixer": "10.4.16",
     "eslint": "8.54.0",
     "eslint-config-next": "14.0.4",
@@ -26,6 +26,6 @@
     "typescript": "5.3.3"
   },
   "devDependencies": {
-    "@uniformdev/cli": "19.130.0"
+    "@uniformdev/cli": "19.161.0"
   }
 }
