+++ examples/nextjs-app-router-localization-starter/src/app/[...path]/page.tsx
