+++ examples/nextjs-app-router-localization-starter/src/app/page.tsx
@@ -1,0 +1,6 @@
+import {redirect} from 'next/navigation';
+
+// This page only renders when the app is built statically (output: 'export')
+export default function RootPage() {
+  redirect('/en');
+}
